Rename io package to IO_dossier
